Move resource-container-helper to core/dom
