+++ src/core/dom/resource-container-helper.js
